+++ test/utils/CompanyRepositoryMock.scala
@@ -22,7 +22,7 @@ class CompanyRepositoryMock(database: mutable.Map[UUID, Company] = mutable.Map.e
       search: CompanyRegisteredSearch,
       paginate: PaginatedSearch,
       user: User
-  ): Future[PaginatedResult[(Company, Int, Int)]] = ???
+  ): Future[PaginatedResult[(Company, Long, Long)]] = ???
 
   override def getOrCreate(siret: SIRET, data: Company): Future[Company] = ???
